integrated bibsnet. categorized functions
