+++ parameter-file-container.json
@@ -1,102 +1,10 @@
 {
     "common": {
-        "fsl_bin_path": "/opt/fsl-6.0.5.1/bin/",
-        "task_id": null
-    },
-
-    "resource_management": {
-        "mem_mb": null,
-        "n_cpus": null,
-        "nipype_plugin_file": null,
-        "nthreads": null,
-        "omp_nthreads": null,
-        "resource_monitor": null
+        "fsl_bin_path": "/opt/fsl-6.0.5.1/bin/"
     },
 
     "bibsnet": {
         "model": "3d_fullres",
-        "nnUNet_predict_path": "/opt/conda/bin/nnUNet_predict",
-        "code_dir": "/home/cabinet/SW/BIBSnet"
-    },
-    
-    "nibabies": {
-        "anat_derivatives": null,
-        "anat_only": null,
-        "aroma_melodic_dimensionality": null,
-        "bids_database_dir": null,
-        "bids_filter_file": null,
-        "boilerplate_only": null,
-        "bold2t1w_dof": null,
-        "bold2t1w_init": null,
-        "cifti_output": "91k",
-        "clean_workdir": null,
-        "config_file": null,
-        "debug": null,
-        "derivatives": "./deriv",
-        "dummy_scans": null,
-        "dvars_spike_threshold": null,
-        "echo_idx": null,
-        "error_on_aroma_warnings": null,
-        "fd_radius": 45,
-        "fd_spike_threshold": 0.2,
-        "fmap_bspline": null,
-        "fmap_no_demean": null,
-        "force_bbr": null,
-        "force_no_bbr": null,
-        "force_syn": null,
-        "fs_license_file": null,
-        "fs_no_reconall": null,
-        "fs_subjects_dir": null,
-        "ignore": null,
-        "longitudinal": null,
-        "medial_surface_nan": null,
-        "nipype_plugin_file": null,
-        "no_submm_recon": null,
-        "notrack": null,
-        "output_layout": null,
-        "output_spaces": null,
-        "random_seed": null,
-        "reports_only": null,
-        "return_all_components": null,
-        "segmentation_atlases_dir": null,
-        "skip_bids_validation": null,
-        "skull_strip_fixed_seed": null,
-        "skull_strip_t1w": null,
-        "skull_strip_template": null,
-        "singularity_image_path": "/home/faird/shared/code/external/pipelines/nibabies/nibabies-22.0.1.sif",
-        "slice_time_ref": null,
-        "sloppy": null,
-        "stop_on_first_crash": null,
-        "use_aroma": null,
-        "use_syn_sdc": null,
-        "work_dir": "work",
-        "write_graph": null
-    },
-
-    "XCPD": {
-        "band_stop_max": null,
-        "band_stop_min": null,
-        "bpf_order": null,
-        "brain_template": null,
-        "cifti": true,
-        "clean_workdir": null,
-        "combineruns": true,
-        "custom_conf": null,
-        "despike": null,
-        "dummytime": null,
-        "fd_thresh": 0.2,
-        "fmri_dir": null,
-        "head_radius": 45,
-        "input_type": "fmirprep",
-        "lower_bpf": null,
-        "motion_filter_order": null,
-        "motion_filter_type": null,
-        "notrack": null,
-        "nuissance_regressors": null,
-        "output_dir": null,
-        "singularity_image_path": null,
-        "smoothing": null,
-        "upper_bpf": null,
-        "work_dir": "work"
+        "nnUNet_predict_path": "/opt/conda/bin/nnUNet_predict"
     }
 }
