+++ src/logger.py
@@ -1,0 +1,16 @@
+import logging
+import sys
+import os
+
+def make_logger():
+    """
+    Make logger to log status updates, warnings, and other important info
+    :return: logging.Logger able to print info to stdout and problems to stderr
+    """  # TODO Incorporate pprint to make printed JSONs/dicts more readable
+    fmt = "\n%(levelname)s %(asctime)s: %(message)s"
+    logging.basicConfig(stream=sys.stdout, format=fmt, level=logging.INFO)  
+    logging.basicConfig(stream=sys.stderr, format=fmt, level=logging.ERROR)
+    logging.basicConfig(stream=sys.stderr, format=fmt, level=logging.WARNING)
+    return logging.getLogger(os.path.basename(sys.argv[0]))
+
+LOGGER = make_logger()
