+++ src/param-types.json
@@ -1,100 +1,11 @@
 {
     "common": {
-        "fsl_bin_path": "existing_directory_path",
-        "task_id": "str"
+        "fsl_bin_path": "existing_directory_path"
     },
 
-    "resource_management": {
-        "mem_mb": "positive_int",
-        "n_cpus": "positive_int",
-        "nipype_plugin_file": "existing_file_path",
-        "nthreads": "positive_int",
-        "omp_nthreads": "positive_int",
-        "resource_monitor": "bool"
-    },
-  
     "bibsnet": {
-        "code_dir": "existing_directory_path",
         "model": "str",
         "nnUNet_predict_path": "existing_file_path" 
-    },
-    
-    "nibabies": {
-        "anat_derivatives": "optional_real_dirpath",
-        "anat_only": "bool",
-        "aroma_melodic_dimensionality": "int",
-        "bids_filter_file": "existing_json_file_path",
-        "boilerplate_only": "bool",
-        "bold2t1w_dof": [6, 9, 12],
-        "bold2t1w_init": ["register", "header"],
-        "cifti_output": ["91k", "170k"],
-        "clean_workdir": "bool",
-        "config_file": "existing_file_path",
-        "debug": ["compcor", "registration", "fieldmaps", "all"],
-        "derivatives": "new_directory_path",
-        "dummy_scans": "positive_int",
-        "dvars_spike_threshold": "positive_float",
-        "echo_idx": "str",
-        "error_on_aroma_warnings": "bool",
-        "fd_radius": "positive_int",
-        "fd_spike_threshold": "float_0_to_1",
-        "fmap_bspline": "bool",
-        "fmap_no_demean": "bool",
-        "force_bbr": "bool",
-        "force_no_bbr": "bool",
-        "force_syn": "bool",
-        "fs_license_file": "existing_file_path",
-        "fs_no_reconall": "bool",
-        "fs_subjects_dir": "optional_real_dirpath",
-        "ignore": ["fieldmaps", "slicetiming", "sbref", "t2w", "flair"],
-        "longitudinal": "bool",
-        "medial_surface_nan": "bool",
-        "no_submm_recon": "bool",
-        "notrack": "bool",
-        "output_layout": ["bids", "legacy"],
-        "output_spaces": "str_list",
-        "random_seed": "bool",
-        "reports_only": "bool",
-        "return_all_components": "bool",
-        "singularity_image_path": "existing_file_path",
-        "segmentation_atlases_dir": "optional_real_dirpath",
-        "skip_bids_validation": "bool",
-        "skull_strip_fixed_seed": "bool",
-        "skull_strip_t1w": ["auto", "skip", "force"],
-        "skull_strip_template": "str",
-        "slice_time_ref": "float_0_to_1",
-        "sloppy": "bool",
-        "stop_on_first_crash": "bool",
-        "use_aroma": "bool",
-        "use_syn_sdc": "bool",
-        "work_dir": "new_directory_path",
-        "write_graph": "bool"
-    },
-
-    "XCPD": {
-        "band_stop_max": "float",
-        "band_stop_min": "float",
-        "bpf_order": "positive_int",
-        "brain_template": "str",
-        "cifti": "bool",
-        "clean_workdir": "bool",
-        "combineruns":"bool",
-        "custom_conf": "existing_tsv_file",
-        "despike": "bool",
-        "dummytime": "positive_float",
-        "fd_thresh": "float_0_to_1",
-        "fmri_dir": "existing_directory_path",
-        "head_radius": "positive_int",
-        "input_type": ["fmirprep", "dcan", "hpc"],
-        "lower_bpf": "float",
-        "motion_filter_order": "positive_int",
-        "motion_filter_type": ["lp", "notch"],
-        "notrack": "bool",
-        "nuissance_regressors": ["27P", "36P", "24P", "acompcor", "tcompcor", "aroma"],
-        "output_dir": "new_directory_path",
-        "singularity_image_path": "existing_file_path",
-        "smoothing": "float",
-        "upper_bpf": "float",
-        "work_dir": "new_directory_path"
     }
+
 }
