+++ src/utilities.py
@@ -8,7 +8,6 @@ Created: 2021-11-12
 Updated: 2023-01-26
 """
 # Import standard libraries
-import argparse
 from datetime import datetime  # for seeing how long scripts take to run
 from glob import glob
 import json
@@ -16,10 +15,11 @@ import nibabel as nib
 from nipype.interfaces import fsl
 import numpy as np
 import os
-import pdb
 import shutil
 import subprocess
 import sys
+
+from src.logger import LOGGER
 
 # Chirality-checking constants
 CHIRALITY_CONST = dict(UNKNOWN=0, LEFT=1, RIGHT=2, BILATERAL=3)
@@ -33,13 +33,12 @@ SCRIPT_DIR = os.path.dirname(os.path.dirname(__file__))
 # NOTE All functions below are in alphabetical order.
 
 
-def align_ACPC_1_img(j_args, logger, xfm_ACPC_vars, crop2full, output_var, t,
+def align_ACPC_1_img(j_args, xfm_ACPC_vars, crop2full, output_var, t,
                      averaged_image):
     """ 
     Functionality copied from the DCAN Infant Pipeline:
     github.com/DCAN-Labs/dcan-infant-pipeline/blob/master/PreFreeSurfer/scripts/ACPCAlignment_with_crop.sh
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :param xfm_ACPC_vars: Dictionary mapping strings (ACPC input arguments'
                           names) to strings (ACPC arguments, file/dir paths)
     :param crop2full: String, valid path to existing crop2full.mat file
@@ -58,46 +57,37 @@ def align_ACPC_1_img(j_args, logger, xfm_ACPC_vars, crop2full, output_var, t,
             for fname in ("crop2acpc", "full2acpc", "full2crop",
                           "acpc2rigidbody")}  # .mat file paths
     
-    run_FSL_sh_script(j_args, logger, "flirt", "-interp", "spline",  
+    run_FSL_sh_script(j_args, "flirt", "-interp", "spline",  
                       "-ref", mni_ref_img_path, "-in", input_img,
                       "-omat", mats["crop2acpc"], # "-out", os.path.join(work_dir, "T{}w_acpc_final.nii.gz".format(t)),
                       "-searchrx", "-45", "45", "-searchry", "-30", "30",
                       "-searchrz", "-30", "30")
 
     # Invert crop2full to get full2crop
-    run_FSL_sh_script(j_args, logger, "convert_xfm", "-inverse", crop2full,
+    run_FSL_sh_script(j_args, "convert_xfm", "-inverse", crop2full,
                       "-omat", mats["full2crop"])  # TODO Move this to right after making crop2full to use it in both T?w-only and here
 
     run_FSL_sh_script(  # Combine ACPC-alignment with robustFOV output
-        j_args, logger, "convert_xfm", "-omat", mats["full2acpc"],
+        j_args, "convert_xfm", "-omat", mats["full2acpc"],
         "-concat", mats["crop2acpc"], mats["full2crop"]
     )
 
     # Transform 12 dof matrix to 6 dof approximation matrix
-    run_FSL_sh_script(j_args, logger, "aff2rigid", mats["full2acpc"],
+    run_FSL_sh_script(j_args, "aff2rigid", mats["full2acpc"],
                       mats["acpc2rigidbody"])
 
     # Apply ACPC alignment to the data
     # Create a resampled image (ACPC aligned) using spline interpolation  # TODO Only run this command in debug mode
     # if j_args["common"]["debug"]:
-    run_FSL_sh_script(j_args, logger, "applywarp", "--rel", "--interp=spline",  
+    run_FSL_sh_script(j_args, "applywarp", "--rel", "--interp=spline",  
                       "-i", averaged_image, "-r", mni_ref_img_path,  # Changed input_img to average_image 2022-06-16
                       "--premat=" + mats["acpc2rigidbody"], "-o", output_img)
     # pdb.set_trace()  # TODO Add "debug" flag?
     return mats
 
 
-def always_true(*_):
-    """
-    This function is useful when e.g. every type except 1 has a corresponding
-    input validation function, because this function is used for that extra 1 
-    :return: True, regardless of what the input arguments are 
-    """
-    return True
-
-
 def apply_final_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs, outputs,
-                         t, j_args, logger):
+                         t, j_args):
     """
     Apply entire image transformation (from cropped image to BIBSnet format)
     with ACPC transformation to a T1w or T2w image 
@@ -108,7 +98,6 @@ def apply_final_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs, outputs,
     :param outputs: Dict that will have T1w &/or T2w ACPC transformed images
     :param t: Int, either 1 or 2 (to signify T1w or T2w respectively)
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :return: outputs, with paths to ACPC-transformed image and transform .mat
     """
     outputs[f"T{t}w"] = get_preBIBS_final_img_fpath_T(
@@ -127,7 +116,7 @@ def apply_final_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs, outputs,
     # space before ACPC and T1w does not 
     if t == 2:
         run_FSL_sh_script( 
-            j_args, logger, "convert_xfm", "-omat", to_rigidbody_final_mat,
+            j_args, "convert_xfm", "-omat", to_rigidbody_final_mat,
             "-concat", xfm_imgs[f"cropT{t}tocropT1"],
             acpc2rigidbody
         )
@@ -137,14 +126,14 @@ def apply_final_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs, outputs,
     if not os.path.exists(crop2BIBS_mat):
         shutil.copy2(to_rigidbody_final_mat, crop2BIBS_mat)
         if j_args["common"]["verbose"]:
-            logger.info("Copying {} to {}".format(to_rigidbody_final_mat,
+            LOGGER.info("Copying {} to {}".format(to_rigidbody_final_mat,
                                                     crop2BIBS_mat))
     outputs[f"T{t}w_crop2BIBS_mat"] = crop2BIBS_mat
 
     # Do the applywarp FSL command from align_ACPC_1_img (for T1w and T2w, for ACPC)
     # applywarp output is optimal_realigned_imgs input
     # Apply registration and ACPC alignment to the T1ws and the T2ws
-    run_FSL_sh_script(j_args, logger, "applywarp", "--rel", 
+    run_FSL_sh_script(j_args, "applywarp", "--rel", 
                         "--interp=spline", "-i", avg_imgs[f"T{t}w_avg"],
                         "-r", xfm_vars["ref_img"].format(t),
                         "--premat=" + crop2BIBS_mat, # preBIBS_ACPC_out["T{}w_crop2BIBS_mat".format(t)],
@@ -155,7 +144,7 @@ def apply_final_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs, outputs,
 
 
 def apply_final_non_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs,
-                             outputs, t, full2crop_ACPC, j_args, logger):
+                             outputs, t, full2crop_ACPC, j_args):
     """
     Apply entire image transformation (from cropped image to BIBSnet format)
     without ACPC transformation to a T1w or T2w image 
@@ -166,7 +155,6 @@ def apply_final_non_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs,
     :param outputs: Dict that will have T1w &/or T2w ACPC transformed images
     :param t: Int, either 1 or 2 (to signify T1w or T2w respectively)
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :return: outputs, with paths to ACPC-transformed image and transform .mat
     """
     outputs[f"T{t}w"] = get_preBIBS_final_img_fpath_T(
@@ -182,13 +170,13 @@ def apply_final_non_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs,
     full2crop_mat = os.path.join(xfm_vars["out_dir"],
                                  f"full2cropT{t}w.mat")
     run_FSL_sh_script( 
-        j_args, logger, "convert_xfm",
+        j_args, "convert_xfm",
         "-omat", full2crop_mat,
         "-concat", full2crop_ACPC, xfm_imgs["cropT1tocropT1"]
     )
     if t == 1:
         run_FSL_sh_script( 
-            j_args, logger, "convert_xfm",
+            j_args, "convert_xfm",
             "-omat", outputs[f"T{t}w_crop2BIBS_mat"],
             "-concat", full2crop_mat, 
             xfm_imgs[f"T{t}w_crop2BIBS_mat"]
@@ -197,10 +185,10 @@ def apply_final_non_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs,
         crop_and_reg_mat = os.path.join(xfm_vars["out_dir"],
                                         "full2cropT2toT1.mat")
         run_FSL_sh_script( 
-            j_args, logger, "convert_xfm", "-omat", crop_and_reg_mat,
+            j_args, "convert_xfm", "-omat", crop_and_reg_mat,
             "-concat", xfm_imgs[f"cropT{t}tocropT1"], full2crop_mat 
         )
-        run_FSL_sh_script(j_args, logger, "convert_xfm", "-omat",
+        run_FSL_sh_script(j_args, "convert_xfm", "-omat",
                           outputs[f"T{t}w_crop2BIBS_mat"], "-concat",
                           xfm_imgs[f"T{t}w_crop2BIBS_mat"], crop_and_reg_mat)
 
@@ -208,7 +196,7 @@ def apply_final_non_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs,
     # (for T2w and not T1w, for non-ACPC)
     # applywarp output is optimal_realigned_imgs input
     # Apply registration to the T1ws and the T2ws
-    run_FSL_sh_script(j_args, logger, "applywarp", "--rel",
+    run_FSL_sh_script(j_args, "applywarp", "--rel",
                       "--interp=spline", "-i", avg_imgs[f"T{t}w_avg"],
                       "-r", xfm_vars["ref_img"].format(t),
                       "--premat=" + outputs[f"T{t}w_crop2BIBS_mat"],
@@ -217,7 +205,7 @@ def apply_final_non_ACPC_xfm(xfm_vars, xfm_imgs, avg_imgs,
 
 
 def apply_final_prebibsnet_xfms(regn_non_ACPC, regn_ACPC, averaged_imgs,
-                                j_args, logger):
+                                j_args):
     """
     Resize the images to match the dimensions of images trained in the model,
     and ensure that the first image (presumably a T1) is co-registered to the
@@ -232,7 +220,6 @@ def apply_final_prebibsnet_xfms(regn_non_ACPC, regn_ACPC, averaged_imgs,
     :param averaged_imgs: Dictionary mapping ints, (T) 1 or 2, to strings
                           (valid paths to existing image files to resize)
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :return: Dict with paths to either ACPC- or non-ACPC image & .mat files
     """
     out_ACPC = dict()
@@ -242,7 +229,7 @@ def apply_final_prebibsnet_xfms(regn_non_ACPC, regn_ACPC, averaged_imgs,
         # Apply ACPC-then-registration transforms for this subject session & T
         out_ACPC.update(apply_final_ACPC_xfm(
             regn_ACPC["vars"], regn_ACPC["img_paths"],
-            averaged_imgs, out_ACPC, t, j_args, logger
+            averaged_imgs, out_ACPC, t, j_args
         ))
 
         # Retrieve path to ACPC full2crop.mat file (to use for non-ACPC xfms)
@@ -251,23 +238,14 @@ def apply_final_prebibsnet_xfms(regn_non_ACPC, regn_ACPC, averaged_imgs,
         # Apply registration-only transforms for this subject session (and T)
         out_non_ACPC.update(apply_final_non_ACPC_xfm(
             regn_non_ACPC["vars"], regn_non_ACPC["img_paths"],
-            averaged_imgs, out_non_ACPC, t, full2crop_ACPC, j_args, logger
+            averaged_imgs, out_non_ACPC, t, full2crop_ACPC, j_args
         ))
 
     # Outputs: 1 .mat file for ACPC and 1 for non-ACPC (only retain the -to-T1w .mat file after this point)
 
     # Return the best of the 2 resized images
     return optimal_realigned_imgs(out_non_ACPC,  # TODO Add 'if' statement to skip eta-squared functionality if T1-/T2-only, b/c only one T means we'll only register to ACPC space
-                                  out_ACPC, j_args, logger)
-
-
-def argify(argname, argval):
-    """
-    :param argname: String naming a parameter for a script called from terminal
-    :param argval: Object to assign in string form as the value of the argument
-    :return: String, a parameter assignment for a script called from terminal
-    """
-    return "--{}={}".format(argname, argval)
+                                  out_ACPC, j_args)
 
 
 def as_cli_arg(arg_str):
@@ -366,7 +344,6 @@ def correct_chirality(nifti_input_file_path, segment_lookup_table,
     :param xfm_ref_img: String, path to (T1w, unless running in T2w-only mode) 
                         image to use as a reference when applying transform
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :return: Dict with paths to native and chirality-corrected images
     """
     nifti_file_paths = dict()
@@ -406,7 +383,7 @@ def correct_chirality(nifti_input_file_path, segment_lookup_table,
     return nifti_file_paths
 
 
-def create_anatomical_averages(avg_params, logger):
+def create_anatomical_averages(avg_params):
     """
     Creates a NIFTI file whose voxels are the average of the voxel values of the input files.
     :param avg_params: Dictionary with 4 keys:
@@ -418,7 +395,7 @@ def create_anatomical_averages(avg_params, logger):
     for t in (1, 2):
         if avg_params.get(f"T{t}w_input"):
             register_and_average_files(avg_params[f"T{t}w_input"],
-                                       avg_params[f"T{t}w_avg"], logger)
+                                       avg_params[f"T{t}w_avg"])
 
 
 def create_avg_image(output_file_path, registered_files):
@@ -448,18 +425,17 @@ def create_avg_image(output_file_path, registered_files):
     nib.save(new_img, output_file_path)
 
 
-def crop_image(input_avg_img, output_crop_img, j_args, logger):
+def crop_image(input_avg_img, output_crop_img, j_args):
     """
     Run robustFOV to crop image
     :param input_avg_img: String, valid path to averaged (T1w or T2w) image
     :param output_crop_img: String, valid path to save cropped image file at
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :return: String, path to crop2full.mat file in same dir as output_crop_img
     """
     output_crop_dir = os.path.dirname(output_crop_img)
     crop2full = os.path.join(output_crop_dir, "crop2full.mat")  # TODO Define this path outside of stages because it's used by preBIBSnet and postBIBSnet
-    run_FSL_sh_script(j_args, logger, "robustfov", "-i", input_avg_img, 
+    run_FSL_sh_script(j_args, "robustfov", "-i", input_avg_img, 
                       "-m", crop2full, "-r", output_crop_img,
                       "-b", j_args["ID"]["brain_z_size"])  # TODO Use head radius for -b
     return crop2full
@@ -646,6 +622,7 @@ def generate_sidecar_json(sub_ses, reference_path, derivs_dir, t, desc):
     with open(file_path, "w+") as file:
         json.dump(sidecar, file)
 
+
 def get_and_make_preBIBSnet_work_dirs(j_args):
     """ 
     :param j_args: Dictionary containing all args from parameter .JSON file
@@ -781,29 +758,6 @@ def get_preBIBS_final_img_fpath_T(t, parent_dir, sub_ses_ID):
     ))
 
 
-def get_spatial_resolution_of(image_fpath, j_args, logger, fn_name="fslinfo"):
-    """
-    :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
-    :param fn_name: String naming the function which is an
-                    executable file in j_args[common][fsl_bin_path]
-    """  # TODO Do we even need this function?
-    # FSL command to run in a subprocess
-    to_run = [os.path.join(j_args["common"]["fsl_bin_path"], fn_name),
-              image_fpath]
-
-    # Run FSL command and read the file information into a dictionary
-    if j_args["common"]["verbose"]:
-        logger.info("Now running FSL command:\n{}"
-                    .format(" ".join(to_run)))
-    img_info = dict()
-    for eachline in subprocess.check_output(to_run).decode("utf-8").split("\n"):
-        split = eachline.split()
-        img_info[split[0]] = split[-1]
-
-    return img_info["pixdim3"]  # A.K.A. brain_z_size
-
-
 def get_stage_name(stage_fn):
     """ 
     :param stage_fn: Function to run one stage of CABINET. Its name must start
@@ -811,18 +765,6 @@ def get_stage_name(stage_fn):
     :return: String naming the CABINET stage to run
     """
     return stage_fn.__name__[4:].lower()
-
-
-def get_sub_base(j_args, run_num=None):
-    """
-    :param j_args: Dictionary containing all args from parameter .JSON file
-    :param run_num: Whole number as an int or string defining which run this is
-    :return: String identifying a subject, session, task, and maybe run
-    """
-    parts = [get_subj_ses(j_args), "task-" + j_args["common"]["task_id"]]
-    if run_num is not None:
-        parts.append(f"run-{run_num}")
-    return "_".join(parts)
 
 
 def get_subj_ID_and_session(j_args):
@@ -835,14 +777,6 @@ def get_subj_ID_and_session(j_args):
     sub = ensure_prefixed(j_args["ID"]["subject"], "sub-")
     return [sub, ensure_prefixed(j_args["ID"]["session"], "ses-")
             ] if dict_has(j_args["ID"], "session") else [sub]
-
-
-def get_subj_ses(j_args):
-    """
-    :param j_args: Dictionary containing all args from parameter .JSON file
-    :return: String combining subject ID and session from parameter file
-    """
-    return "_".join(get_subj_ID_and_session(j_args))
 
 
 def get_template_age_closest_to(age, templates_dir):
@@ -880,20 +814,9 @@ def get_age_closest_to(subject_age, all_ages):
     :return: Int, the age in all_ages which is closest to subject_age
     """
     return all_ages[np.argmin(np.abs(np.array(all_ages)-subject_age))]
-    
-
-def glob_and_copy(dest_dirpath, *path_parts_to_glob):
-    """
-    Collect all files matching a glob string, then copy those files
-    :param dest_dirpath: String, a valid path of a directory to copy files into
-    :param path_parts_to_glob: Unpacked list of strings which join to form a
-                               glob string of a path to copy files from
-    """
-    for file_src in glob(os.path.join(*path_parts_to_glob)):
-        shutil.copy(file_src, dest_dirpath)
 
 
-def log_stage_finished(stage_name, event_time, sub_ses, logger):
+def log_stage_finished(stage_name, event_time, sub_ses):
     """
     Print and return a string showing how much time has passed since the
     current running script reached a certain part of its process
@@ -903,7 +826,7 @@ def log_stage_finished(stage_name, event_time, sub_ses, logger):
     :return: String with an easily human-readable message showing how much time
              has passed since {stage_start} when {stage_name} started.
     """
-    logger.info("{0} finished on subject {1}. "
+    LOGGER.info("{0} finished on subject {1}. "
                 "Time elapsed since {0} started: {2}"
                 .format(stage_name, " session ".join(sub_ses),
                         datetime.now() - event_time))
@@ -934,35 +857,34 @@ def only_Ts_needed_for_bibsnet_model(sub_ses_ID):
             yield t
 
 
-def optimal_realigned_imgs(xfm_imgs_non_ACPC, xfm_imgs_ACPC_and_reg, j_args, logger):
+def optimal_realigned_imgs(xfm_imgs_non_ACPC, xfm_imgs_ACPC_and_reg, j_args):
     """
     Check whether the cost function shows that only the registration-T2-to-T1
     or the ACPC-alignment-and-T2-to-T1-registration is better (check whether
     ACPC alignment improves the T2-to-T1 registration; compare the T2-to-T1
     with and without first doing the ACPC registration)
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     """
     msg = "Using {} T2w-to-T1w registration for resizing.\nT1w: {}\nT2w: {}"
     eta = dict()
-    logger.info("\nACPC:")
+    LOGGER.info("\nACPC:")
     eta["ACPC"] = calculate_eta(xfm_imgs_ACPC_and_reg)
-    logger.info("\nNon-ACPC:")
+    LOGGER.info("\nNon-ACPC:")
     eta["non-ACPC"] = calculate_eta(xfm_imgs_non_ACPC)
-    logger.info(f"Eta-Squared Values: {eta}")
+    LOGGER.info(f"Eta-Squared Values: {eta}")
     if eta["non-ACPC"] > eta["ACPC"]:
         optimal_resize = xfm_imgs_non_ACPC
-        logger.info(msg.format("only", optimal_resize["T1w"],
+        LOGGER.info(msg.format("only", optimal_resize["T1w"],
                                optimal_resize["T2w"]))  # TODO Verify that these print the absolute path
     else:
         optimal_resize = xfm_imgs_ACPC_and_reg
-        logger.info(msg.format("ACPC and", optimal_resize["T1w"],
+        LOGGER.info(msg.format("ACPC and", optimal_resize["T1w"],
                                optimal_resize["T2w"]))  # TODO Verify that these print the absolute path
     return optimal_resize
 
 
 def register_preBIBSnet_imgs_ACPC(cropped_imgs, output_dir, xfm_non_ACPC_vars,
-                                  crop2full, averaged_imgs, j_args, logger):
+                                  crop2full, averaged_imgs, j_args):
     """
     :param cropped_imgs: Dictionary mapping ints, (T) 1 or 2, to strings (valid
                          paths to existing image files to resize)
@@ -972,7 +894,6 @@ def register_preBIBSnet_imgs_ACPC(cropped_imgs, output_dir, xfm_non_ACPC_vars,
     :param averaged_imgs: Dictionary mapping ints, (T) 1 or 2, to strings
                           (valid paths to existing image files to resize)
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     """
     # Build dict of variables used for image transformation with ACPC alignment
     xfm_ACPC_vars = xfm_non_ACPC_vars.copy()
@@ -1001,13 +922,13 @@ def register_preBIBSnet_imgs_ACPC(cropped_imgs, output_dir, xfm_non_ACPC_vars,
 
         # Run ACPC alignment
         xfm_ACPC_vars[f"mats_T{t}w"] = align_ACPC_1_img(
-            j_args, logger, xfm_ACPC_vars, crop2full[t], reg_in_var, t,
+            j_args, xfm_ACPC_vars, crop2full[t], reg_in_var, t,
             averaged_imgs[f"T{t}w_avg"]
         )
 
     # T2w-T1w alignment of ACPC-aligned images
     xfm_ACPC_and_reg_imgs = registration_T2w_to_T1w(
-        j_args, logger, xfm_ACPC_vars, reg_in_var, acpc=True
+        j_args, xfm_ACPC_vars, reg_in_var, acpc=True
     )
 
     # pdb.set_trace()  # TODO Add "debug" flag?
@@ -1016,7 +937,7 @@ def register_preBIBSnet_imgs_ACPC(cropped_imgs, output_dir, xfm_non_ACPC_vars,
 
 
 def register_preBIBSnet_imgs_non_ACPC(cropped_imgs, output_dir, ref_image, 
-                                      ident_mx, resolution, j_args, logger):
+                                      ident_mx, resolution, j_args):
     """
     :param cropped_imgs: Dictionary mapping ints, (T) 1 or 2, to strings (valid
                          paths to existing image files to resize)
@@ -1028,7 +949,6 @@ def register_preBIBSnet_imgs_non_ACPC(cropped_imgs, output_dir, ref_image,
     :param ident_mx: String, valid path to existing identity matrix .mat file
     :param resolution:
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     """
     # TODO Add 'if' to skip most of the functionality here for T1-only or T2-only
 
@@ -1057,7 +977,7 @@ def register_preBIBSnet_imgs_non_ACPC(cropped_imgs, output_dir, ref_image,
     os.makedirs(xfm_non_ACPC_vars["out_dir"], exist_ok=True)
 
     xfm_imgs_non_ACPC = registration_T2w_to_T1w(
-        j_args, logger, xfm_non_ACPC_vars, reg_in_var, acpc=False
+        j_args, xfm_non_ACPC_vars, reg_in_var, acpc=False
     )
 
     # pdb.set_trace()  # TODO Add "debug" flag?
@@ -1065,7 +985,7 @@ def register_preBIBSnet_imgs_non_ACPC(cropped_imgs, output_dir, ref_image,
     return {"vars": xfm_non_ACPC_vars, "img_paths": xfm_imgs_non_ACPC}
 
 
-def register_and_average_files(input_file_paths, output_file_path, logger):
+def register_and_average_files(input_file_paths, output_file_path):
     """
     Register all input image files, and if there are multiple of them, then
     create an average of all of them
@@ -1073,27 +993,25 @@ def register_and_average_files(input_file_paths, output_file_path, logger):
                              image file to register
     :param output_file_path: String, valid path to image file to create by
                              averaging all of the input_file_paths images
-    :param logger: logging.Logger object to show messages and raise warnings
     """
     reference = input_file_paths[0]
     out_dir=os.path.dirname(output_file_path)
     if len(input_file_paths) > 1:
         registered_files = register_files(input_file_paths, reference,
-                                          out_dir, logger)
+                                          out_dir)
 
         create_avg_image(output_file_path, registered_files)
     else:
         shutil.copyfile(reference, output_file_path)
 
 
-def register_files(input_file_paths, reference, out_dir, logger):
+def register_files(input_file_paths, reference, out_dir):
     """
     :param input_file_paths: List of strings, each a valid path to an existing
                              image file to register
     :param reference: String, valid path to existing img to register others to
     :param out_dir: String, valid path to existing directory to save registered
                     images into
-    :param logger: logging.Logger object to show messages and raise warnings
     :raises RuntimeError: If FSL FLIRT command to register images fails
     :return: List of strings, each a valid path to a newly created image file,
              starting with the reference image and then every input_file_paths
@@ -1116,7 +1034,7 @@ def register_files(input_file_paths, reference, out_dir, logger):
 
         # Build FSL command to register each file
         flt.inputs.in_file = structural
-        logger.info("Now running FSL FLIRT:\n{}".format(flt.cmdline))
+        LOGGER.info("Now running FSL FLIRT:\n{}".format(flt.cmdline))
         out_index = flt.cmdline.find('-out')
         start_index = out_index + len('-out') + 1
         end_index = flt.cmdline.find(' ', start_index)
@@ -1132,11 +1050,10 @@ def register_files(input_file_paths, reference, out_dir, logger):
     return registered_files
     
 
-def registration_T2w_to_T1w(j_args, logger, xfm_vars, reg_input_var, acpc):
+def registration_T2w_to_T1w(j_args, xfm_vars, reg_input_var, acpc):
     """
     T2w to T1w registration for use in preBIBSnet
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :param xfm_vars: Dictionary containing paths to files used in registration
     :param reg_input_var: String naming the key in xfm_vars mapped to the path
                           to the image to use as an input for registration
@@ -1147,7 +1064,7 @@ def registration_T2w_to_T1w(j_args, logger, xfm_vars, reg_input_var, acpc):
     # to the image to use as an input for registration
     inputs_msg = "\n".join(["T{}w: {}".format(t, xfm_vars[reg_input_var.format(t)])
                             for t in only_Ts_needed_for_bibsnet_model(j_args["ID"])])
-    logger.info("Input images for T1w registration:\n" + inputs_msg)
+    LOGGER.info("Input images for T1w registration:\n" + inputs_msg)
 
     # Define paths to registration output matrices and images
     registration_outputs = {"cropT1tocropT1": xfm_vars["ident_mx"],
@@ -1174,7 +1091,7 @@ def registration_T2w_to_T1w(j_args, logger, xfm_vars, reg_input_var, acpc):
         )
 
         if t == 2:  # Make T2w-to-T1w matrix
-            run_FSL_sh_script(j_args, logger, "flirt",
+            run_FSL_sh_script(j_args, "flirt",
                             "-ref", xfm_vars[reg_input_var.format(1)],
                             "-in", xfm_vars[reg_input_var.format(2)],
                             "-omat", registration_outputs["cropT2tocropT1"],
@@ -1192,10 +1109,10 @@ def registration_T2w_to_T1w(j_args, logger, xfm_vars, reg_input_var, acpc):
             transform_image_T(
                 t, (xfm_vars[reg_input_var.format(t)] if t == 1 else
                     registration_outputs["T2w"]),
-                xfm_vars, registration_outputs, j_args, logger
+                xfm_vars, registration_outputs, j_args
             )
             run_FSL_sh_script(  # TODO Should the output image even be created here, or during applywarp?
-                j_args, logger, "flirt",
+                j_args, "flirt",
                 "-in", xfm_vars[reg_input_var.format(t)] if t == 1 else registration_outputs["T2w"],  # Input: Cropped image
                 "-ref", xfm_vars["ref_img"].format(t),
                 "-applyisoxfm", xfm_vars["resolution"],
@@ -1207,7 +1124,7 @@ def registration_T2w_to_T1w(j_args, logger, xfm_vars, reg_input_var, acpc):
     return registration_outputs
 
 
-def transform_image_T(t, cropped_in_img, xfm_vars, regn_outs, j_args, logger):
+def transform_image_T(t, cropped_in_img, xfm_vars, regn_outs, j_args):
     """
     Run FSL command on a cropped input image to apply a .mat file transform 
     :param t: Int, either 1 or 2 (to signify T1w or T2w respectively)
@@ -1215,10 +1132,9 @@ def transform_image_T(t, cropped_in_img, xfm_vars, regn_outs, j_args, logger):
     :param xfm_vars: Dict with paths to reference image & identity matrix files
     :param regn_outs: Dict with paths to transformed output images to make
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     """
     run_FSL_sh_script(  # TODO Should the output image even be created here, or during applywarp?
-        j_args, logger, "flirt",
+        j_args, "flirt",
         "-in", cropped_in_img, # xfm_vars[reg_input_var.format(t)] if t == 1 else registration_outputs["T2w"],  # Input: Cropped image
         "-ref", xfm_vars["ref_img"].format(t),
         "-applyisoxfm", xfm_vars["resolution"],
@@ -1238,7 +1154,7 @@ def reshape_volume_to_array(array_img):
 
 
 def reverse_regn_revert_to_native(nifti_file_paths, chiral_out_dir,
-                                  xfm_ref_img, t, j_args, logger):
+                                  xfm_ref_img, t, j_args):
     """
     :param nifti_file_paths: Dict with valid paths to native and
                              chirality-corrected images
@@ -1248,7 +1164,6 @@ def reverse_regn_revert_to_native(nifti_file_paths, chiral_out_dir,
                         image to use as a reference when applying transform
     :param t: 1 or 2, whether running on T1 or T2
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :return: String, valid path to existing image reverted to native
     """
     sub_ses = get_subj_ID_and_session(j_args)
@@ -1264,23 +1179,22 @@ def reverse_regn_revert_to_native(nifti_file_paths, chiral_out_dir,
         f"preBIBSnet_*crop_T{t}w_to_BIBS_template.mat"  # TODO Name this outside of pre- and postBIBSnet then pass it to both
     )
     preBIBSnet_mat = glob(preBIBSnet_mat_glob).pop()
-    run_FSL_sh_script(j_args, logger, "convert_xfm", "-omat",
+    run_FSL_sh_script(j_args, "convert_xfm", "-omat",
                       seg2native, "-inverse", preBIBSnet_mat)
     # TODO Define preBIBSnet_mat path outside of stages because it's used by preBIBSnet and postBIBSnet
 
-    run_FSL_sh_script(j_args, logger, "flirt", "-applyxfm",
+    run_FSL_sh_script(j_args, "flirt", "-applyxfm",
                       "-ref", xfm_ref_img, "-in", dummy_copy,
                       "-init", seg2native, "-o", nifti_file_paths[f"native-T{t}"],
                       "-interp", "nearestneighbour")
     return nifti_file_paths[f"native-T{t}"]
 
 
-def run_FSL_sh_script(j_args, logger, fsl_fn_name, *fsl_args):
+def run_FSL_sh_script(j_args, fsl_fn_name, *fsl_args):
     """
     Run any FSL function in a Bash subprocess, unless its outputs exist and the
     parameter file said not to overwrite outputs
     :param j_args: Dictionary containing all args from parameter .JSON file
-    :param logger: logging.Logger object to show messages and raise warnings
     :param fsl_fn_name: String naming the FSL function which is an
                         executable file in j_args[common][fsl_bin_path]
     """
@@ -1302,14 +1216,14 @@ def run_FSL_sh_script(j_args, logger, fsl_fn_name, *fsl_args):
             skip_cmd = True
     if skip_cmd:
         if j_args["common"]["verbose"]:
-            logger.info("Skipping FSL {} command because its output image(s) "
+            LOGGER.info("Skipping FSL {} command because its output image(s) "
                         "listed below exist(s) and overwrite=False.\n{}"
                         .format(fsl_fn_name, "\n".join(outputs)))
 
     # Otherwise, just run the FSL command
     else:
         if j_args["common"]["verbose"]:
-            logger.info("Now running FSL command:\n{}"
+            LOGGER.info("Now running FSL command:\n{}"
                         .format(" ".join(to_run)))
         subprocess.check_call(to_run)
 
@@ -1317,7 +1231,7 @@ def run_FSL_sh_script(j_args, logger, fsl_fn_name, *fsl_args):
 
 
 def run_all_stages(all_stages, sub_ses_IDs, start, end,
-                   ubiquitous_j_args, logger):
+                   ubiquitous_j_args):
     """
     Run stages sequentially, starting and ending at stages specified by user
     :param all_stages: List of functions in order where each runs one stage
@@ -1326,10 +1240,9 @@ def run_all_stages(all_stages, sub_ses_IDs, start, end,
     :param start: String naming the first stage the user wants to run
     :param end: String naming the last stage the user wants to run
     :param ubiquitous_j_args: Dictionary of all args needed by each stage
-    :param logger: logging.Logger object to show messages and raise warnings
     """
     if ubiquitous_j_args["common"]["verbose"]:
-        logger.info("All parameters from input args and input .JSON file:\n{}"
+        LOGGER.info("All parameters from input args and input .JSON file:\n{}"
                     .format(ubiquitous_j_args))
 
     # For every session of every subject...
@@ -1345,7 +1258,7 @@ def run_all_stages(all_stages, sub_ses_IDs, start, end,
         )
 
         # ...check that all required input files exist for the stages to run
-        verify_CABINET_inputs_exist(sub_ses, sub_ses_j_args, logger)
+        verify_CABINET_inputs_exist(sub_ses, sub_ses_j_args)
 
         # ...run all stages that the user said to run
         for stage in all_stages:
@@ -1355,10 +1268,10 @@ def run_all_stages(all_stages, sub_ses_IDs, start, end,
             if running:
                 stage_start = datetime.now()
                 if sub_ses_j_args["common"]["verbose"]:
-                    logger.info("Now running {} stage on:\n{}"
+                    LOGGER.info("Now running {} stage on:\n{}"
                                 .format(name, sub_ses_j_args["ID"]))
-                sub_ses_j_args = stage(sub_ses_j_args, logger)
-                log_stage_finished(name, stage_start, sub_ses, logger)
+                sub_ses_j_args = stage(sub_ses_j_args)
+                log_stage_finished(name, stage_start, sub_ses)
             if name == end:
                 running = False
 
@@ -1384,260 +1297,9 @@ def sum_of_2_sums_of_squares_of(np_vector1, np_vector2, a_mean):
     for each_vec in (np_vector1, np_vector2):
         total_sum += sum(np.square(each_vec - a_mean))
     return total_sum
-                    
-
-def valid_float_0_to_1(val):
-    """
-    :param val: Object to check, then throw an error if it is invalid
-    :return: val if it is a float between 0 and 1 (otherwise invalid)
-    """
-    return validate(val, lambda x: 0 <= float(x) <= 1, float,
-                    "{} is not a number between 0 and 1")
 
 
-def valid_output_dir(path):
-    """
-    Try to make a folder for new files at path; throw exception if that fails
-    :param path: String which is a valid (not necessarily real) folder path
-    :return: String which is a validated absolute path to real writeable folder
-    """
-    return validate(path, lambda x: os.access(x, os.W_OK),
-                    valid_readable_dir, "Cannot create directory at {}",
-                    lambda y: os.makedirs(y, exist_ok=True))
-
-
-def valid_output_dir_or_none(path):
-    """
-    Try to make a folder for new files at path, unless "path" is just False.
-    Throw exception if that fails
-    :param path: String which should be either a valid (not necessarily real)
-                 folder path or None
-    :return: Either None or a validated absolute path to real writeable folder
-    """
-    return path if path is None else valid_output_dir(path)
-
-
-def valid_positive_float(to_validate):
-    """
-    Throw argparse exception unless to_validate is a positive float
-    :param to_validate: Object to test whether it is a positive float
-    :return: to_validate if it is a positive float
-    """
-    return validate(to_validate, lambda x: float(x) >= 0, float,
-                    "{} is not a positive number")
-
-
-def valid_readable_dir(path):
-    """
-    :param path: Parameter to check if it represents a valid directory path
-    :return: String representing a valid directory path
-    """
-    return validate(path, os.path.isdir, valid_readable_file,
-                    "Cannot read directory at '{}'")
-
-
-def valid_readable_file(path):
-    """
-    Throw exception unless parameter is a valid readable filepath string. Use
-    this, not argparse.FileType("r") which leaves an open file handle.
-    :param path: Parameter to check if it represents a valid filepath
-    :return: String representing a valid filepath
-    """
-    return validate(path, lambda x: os.access(x, os.R_OK),
-                    os.path.abspath, "Cannot read file at '{}'")
-
-
-def valid_readable_json(path):
-    """
-    :param path: Parameter to check if it represents a valid .json file path
-    :return: String representing a valid .json file path
-    """
-    return validate(path, lambda _: os.path.splitext(path)[-1] == ".json",
-                    valid_readable_file,
-                    "'{}' is not a path to a readable .json file")
-
-
-def valid_subj_ses(in_arg, pfx, name): 
-    """
-    :param in_arg: Object to check if it is a valid subject ID or session name
-    :param pfx: String that's the prefix to an ID; "sub-" or "ses-"
-    :param name: String describing what in_arg should be (e.g. "subject")
-    :return: True if in_arg is a valid subject ID or session name; else False
-    """
-    return validate(in_arg, always_true, lambda x: ensure_prefixed(x, pfx),
-                    "'{}'" + " is not a valid {}".format(name))
-
-
-def valid_template_filename(fname):
-    """
-    :param fname: Parameter to check if it represents a .fsf file name
-    :return: String representing the .fsf file name
-    """
-    return validate(fname, lambda x: os.path.splitext(x)[-1] == ".fsf",
-                    lambda y: y, "'{}' is not an .fsf file name")
-
-
-# TODO Use --clean-env flag to prevent contamination of any Singularity run by outside environment variables
-#   https://3.basecamp.com/5032058/buckets/21517584/messages/4545156874
-
-
-def valid_time_str(in_arg):
-    """
-    :param in_arg: Object to check if it's a time string in the HH:MM:SS format
-    :return: True if in_arg is a time limit string in that format; else False
-    """
-    try:
-        split = in_arg.split(":")
-        assert len(split) == 3
-        for each_num in split:
-            assert each_num.isdigit()
-            assert int(each_num) >= 0
-        return in_arg
-    except (TypeError, AssertionError, ValueError):
-        raise argparse.ArgumentTypeError("'{}' is not a valid time string"
-                                         .format(in_arg))
-
-
-def valid_whole_number(to_validate):
-    """
-    Throw argparse exception unless to_validate is a positive integer
-    :param to_validate: Object to test whether it is a positive integer
-    :return: to_validate if it is a positive integer
-    """
-    return validate(to_validate, lambda x: int(x) >= 0, int,
-                    "{} is not a positive integer")
-
-
-def valid_subj_ses_ID(to_validate):
-    """
-    :param to_validate: Object to turn into a valid subject/session ID label
-    :return: String, valid subject/session ID label
-    """  # TODO Validate that subject/session exists 
-    return validate(to_validate, always_true, lambda x: x.split("-")[-1],
-                    "{} is not a valid subject/session ID.")
-
-
-def validate(to_validate, is_real, make_valid, err_msg, prepare=None):
-    """
-    Parent/base function used by different type validation functions. Raises an
-    argparse.ArgumentTypeError if the input object is somehow invalid.
-    :param to_validate: String to check if it represents a valid object 
-    :param is_real: Function which returns true iff to_validate is real
-    :param make_valid: Function which returns a fully validated object
-    :param err_msg: String to show to user to tell them what is invalid
-    :param prepare: Function to run before validation
-    :return: to_validate, but fully validated
-    """
-    try:
-        if prepare:
-            prepare(to_validate)
-        assert is_real(to_validate)
-        return make_valid(to_validate)
-    except (OSError, TypeError, AssertionError, ValueError,
-            argparse.ArgumentTypeError):
-        raise argparse.ArgumentTypeError(err_msg.format(to_validate))
-
-
-def validate_parameter_types(j_args, j_types, param_json, parser, stage_names):
-    """
-    Verify that every parameter in j_args is the correct data-type and the 
-    right kind of value. If any parameter is invalid, crash with an error.
-    :param j_args: Dictionary containing all args from parameter .JSON file
-    :param j_types: Dictionary mapping every argument in j_args to its type
-    :param param_json: String, path to readable .JSON file with all parameters
-    :param parser: argparse.ArgumentParser to raise error if anything's invalid
-    :param stage_names: List of strings; each names a stage to run
-    """
-    # Define functions to validate arguments of each data type
-    type_validators = {"bool": bool, "int": int,
-                       "existing_directory_path": valid_readable_dir,
-                       "existing_file_path": valid_readable_file,
-                       "existing_json_file_path": valid_readable_json,
-                       "float_0_to_1": valid_float_0_to_1,
-                       "new_directory_path": valid_output_dir,
-                       "new_file_path": always_true,  # TODO Make "valid_output_filename" function to ensure that filenames don't have spaces or slashes, and maaaaybe to ensure that the new file's parent directory exists?
-                       "optional_new_dirpath": valid_output_dir_or_none,
-                       "optional_real_dirpath": valid_output_dir_or_none,
-                       "positive_float": valid_positive_float,
-                       "positive_int": valid_whole_number, 
-                       "str": always_true}
-
-    required_for_stage = {
-        "nibabies": ["cifti_output", "fd_radius", "work_dir"],
-        "xcpd": ["cifti", "combineruns", "fd_thresh",
-                 "head_radius", "input_type"]
-    }
-
-    # Get a list of all stages after the last stage to run
-    after_end = stage_names[stage_names.index(j_args["stage_names"]["end"])+1:]
-
-    # Verify parameters in each section
-    to_delete = list()
-    for section_orig_name, section_dict in j_types.items():
-        section_name = section_orig_name.lower()  # TODO Should we change parameter types .JSON file to make section names already lowercase?
-
-        # Skip the j_args sections for stages not being run
-        if section_name in stage_names and section_name in after_end:
-            if section_orig_name in j_args:
-                to_delete.append(section_orig_name)
-
-        # Only include resource_management if we're in SLURM/SBATCH job(s)
-        elif not (section_name == "resource_management"
-                  and not j_args["meta"]["slurm"]):
-
-            # Validate every parameter in the section
-            for arg_name, arg_type in section_dict.items():
-
-                # Ignore XCP-D and nibabies parameters that are null
-                arg_value = j_args[section_name][arg_name]
-                if not (arg_value is None and
-                        section_name in ("nibabies", "XCPD") and
-                        arg_value not in required_for_stage[section_name]):
-                    validate_1_parameter(j_args, arg_name, arg_type, section_name,
-                                         type_validators, param_json, parser)
-
-    # Remove irrelevant parameters
-    for section_name in to_delete:
-        del j_args[section_name]
-    return j_args
-
-
-def validate_1_parameter(j_args, arg_name, arg_type, section_name,
-                         type_validators, param_json, parser):
-    """
-    :param j_args: Dictionary containing all args from parameter .JSON file
-    :param arg_name: String naming the parameter to validate
-    :param arg_type: Either a string naming the data type of the parameter to 
-                     validate or a list of options the parameter must be in
-    :param section_name: String that's a subcategory in the param_json file
-    :param type_validators: Dict mapping each arg_type to a validator function
-    :param param_json: String, path to readable .JSON file with all parameters
-    :param parser: argparse.ArgumentParser to raise error if anything's invalid
-    """
-    to_validate = j_args[section_name][arg_name]  # Parameter to validate
-    err_msg = ("'{}' is not a valid '{}' parameter in the '{}' section "
-               "of {} (Problem: {})")  # Message for if to_validate is invalid
-    try:
-        # Run a type validation function unless arg_type is a list
-        if isinstance(arg_type, str):
-            type_validators[arg_type](to_validate)
-            
-        # Verify that the parameter is a valid member of a choices list
-        elif isinstance(arg_type, list) and to_validate not in arg_type:
-            parser.error(
-                err_msg.format(to_validate, arg_name, section_name, param_json,
-                               "Valid {} values: {}"
-                               .format(arg_name, ", ".join(arg_type)))
-            )
-
-    # If type validation fails then inform the user which parameter
-    # has an invalid type and what the valid types are
-    except (argparse.ArgumentTypeError, KeyError, TypeError, ValueError) as e:
-        parser.error(err_msg.format(to_validate, arg_name,
-                                    section_name, param_json, e))
-
-
-def verify_CABINET_inputs_exist(sub_ses, j_args, logger):
+def verify_CABINET_inputs_exist(sub_ses, j_args):
     """
     Given a stage, verify that all of the necessary inputs for that stage exist 
     :param a_stage: String naming a stage
@@ -1674,22 +1336,10 @@ def verify_CABINET_inputs_exist(sub_ses, j_args, logger):
             if not glob(globbable):
                 missing_files.append(globbable)
         if missing_files:
-            logger.error("The file(s) below are needed to run the {} stage, "
+            LOGGER.error("The file(s) below are needed to run the {} stage, "
                         "but they do not exist.\n{}\n"
                         .format(stage, "\n".join(missing_files)))
             sys.exit(1)
 
-    logger.info("All required input files exist.")
+    LOGGER.info("All required input files exist.")
 
-
-def will_run_stage(a_stage, start_stage, end_stage, all_stage_names):
-    """
-    :param a_stage: String naming a stage to run
-    :param start_stage: String naming the first stage to run
-    :param end_stage: String naming the last stage to run
-    :param all_stage_names: List of strings; each names a stage to run
-    :return: True if a_stage is between start_stage and end_stage; else False
-    """
-    return (all_stage_names.index(start_stage)
-            <= all_stage_names.index(a_stage)
-            <= all_stage_names.index(end_stage))
